+++ internal/common/basic.go
@@ -5,16 +5,15 @@ import (
 	"fmt"
 	"time"
 
-	corev1 "k8s.io/api/core/v1"
-	capi "sigs.k8s.io/cluster-api/api/v1beta1"
-	capiexp "sigs.k8s.io/cluster-api/exp/api/v1beta1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	capi "sigs.k8s.io/cluster-api/api/core/v1beta2"
 	cr "sigs.k8s.io/controller-runtime/pkg/client"
 
-	"github.com/giantswarm/clustertest/v3/pkg/application"
-	"github.com/giantswarm/clustertest/v3/pkg/client"
-	"github.com/giantswarm/clustertest/v3/pkg/failurehandler"
-	"github.com/giantswarm/clustertest/v3/pkg/logger"
-	"github.com/giantswarm/clustertest/v3/pkg/wait"
+	"github.com/giantswarm/clustertest/v4/pkg/application"
+	"github.com/giantswarm/clustertest/v4/pkg/client"
+	"github.com/giantswarm/clustertest/v4/pkg/failurehandler"
+	"github.com/giantswarm/clustertest/v4/pkg/logger"
+	"github.com/giantswarm/clustertest/v4/pkg/wait"
 
 	"github.com/giantswarm/cluster-test-suites/v5/internal/state"
 	"github.com/giantswarm/cluster-test-suites/v5/internal/timeout"
@@ -196,13 +195,13 @@ func runBasic() {
 				)
 		})
 
-		It("has Cluster Ready condition with Status='True'", func() {
+		It("has Cluster Available condition with Status='True'", func() {
 			// Overriding the default timeout, when ClusterReadyTimeout is set
 			timeout := state.GetTestTimeout(timeout.ClusterReadyTimeout, 15*time.Minute)
 
 			mcClient := state.GetFramework().MC()
 			cluster := state.GetCluster()
-			Eventually(wait.IsClusterConditionSet(state.GetContext(), mcClient, cluster.Name, cluster.GetNamespace(), capi.ReadyCondition, corev1.ConditionTrue, "")).
+			Eventually(wait.IsClusterConditionSet(state.GetContext(), mcClient, cluster.Name, cluster.GetNamespace(), capi.AvailableCondition, metav1.ConditionTrue, "")).
 				WithTimeout(timeout).
 				WithPolling(wait.DefaultInterval).
 				Should(BeTrue(), failurehandler.LLMPrompt(state.GetFramework(), state.GetCluster(), "Investigate Cluster API Cluster CR not ready"))
@@ -270,11 +269,10 @@ func CheckWorkerNodesReady(ctx context.Context, wcClient *client.Client, values 
 	}
 }
 
-// CheckMachinePoolsReadyAndRunning checks if all MachinePool resources have Ready condition with Status True and are in
-// Running phase.
+// CheckMachinePoolsReadyAndRunning checks if all MachinePool resources are in Running phase with all replicas available.
 func CheckMachinePoolsReadyAndRunning(ctx context.Context, mcClient *client.Client, clusterName string, clusterNamespace string) func() error {
 	return func() error {
-		machinePools := &capiexp.MachinePoolList{}
+		machinePools := &capi.MachinePoolList{}
 		machinePoolListOptions := []cr.ListOption{
 			cr.InNamespace(clusterNamespace),
 			cr.MatchingLabels{
@@ -291,27 +289,30 @@ func CheckMachinePoolsReadyAndRunning(ctx context.Context, mcClient *client.Clie
 			return nil
 		}
 
-		allMachinePoolsAreReadyAndRunning := true
+		allReady := true
 		for _, mp := range machinePools.Items {
-			machinePool := mp
-			var machinePoolIsReady bool
-			machinePoolIsReady, err = wait.IsClusterApiObjectConditionSet(&mp, capi.ReadyCondition, corev1.ConditionTrue, "")
-			if err != nil {
-				return err
-			}
-			allMachinePoolsAreReadyAndRunning = allMachinePoolsAreReadyAndRunning && machinePoolIsReady
+			phase := capi.MachinePoolPhase(mp.Status.Phase)
+			isRunning := phase == capi.MachinePoolPhaseRunning
 
-			currentMachinePoolPhase := capiexp.MachinePoolPhase(machinePool.Status.Phase)
-			machinePoolIsRunning := currentMachinePoolPhase == capiexp.MachinePoolPhaseRunning
-			allMachinePoolsAreReadyAndRunning = allMachinePoolsAreReadyAndRunning && machinePoolIsRunning
+			var desired, available int32
+			if mp.Spec.Replicas != nil {
+				desired = *mp.Spec.Replicas
+			}
+			if mp.Status.AvailableReplicas != nil {
+				available = *mp.Status.AvailableReplicas
+			}
+			replicasReady := available >= desired && desired > 0
+
 			logger.Log(
-				"MachinePool '%s/%s' expected to be in Running phase, found MachinePool is in '%s' phase.",
-				machinePool.Namespace,
-				machinePool.Name,
-				machinePool.Status.Phase)
+				"MachinePool '%s/%s': phase=%s, replicas=%d/%d available",
+				mp.Namespace, mp.Name, mp.Status.Phase, available, desired)
+
+			if !isRunning || !replicasReady {
+				allReady = false
+			}
 		}
 
-		if !allMachinePoolsAreReadyAndRunning {
+		if !allReady {
 			return fmt.Errorf("not all MachinePools are ready and running")
 		}