+++ internal/state/state.go
@@ -7,8 +7,8 @@ import (
 
 	"github.com/giantswarm/cluster-test-suites/v5/internal/timeout"
 
-	"github.com/giantswarm/clustertest/v3"
-	"github.com/giantswarm/clustertest/v3/pkg/application"
+	"github.com/giantswarm/clustertest/v4"
+	"github.com/giantswarm/clustertest/v4/pkg/application"
 )
 
 var lock = &sync.Mutex{}