+++ internal/suite/setup.go
@@ -8,23 +8,23 @@ import (
 	"strings"
 	"time"
 
-	"github.com/giantswarm/clustertest/v3/pkg/failurehandler"
+	"github.com/giantswarm/clustertest/v4/pkg/failurehandler"
 	. "github.com/onsi/ginkgo/v2" //nolint:staticcheck
 	. "github.com/onsi/gomega"    //nolint:staticcheck
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/utils/ptr"
 
-	cb "github.com/giantswarm/cluster-standup-teardown/v4/pkg/clusterbuilder"
-	"github.com/giantswarm/cluster-standup-teardown/v4/pkg/standup"
-	"github.com/giantswarm/clustertest/v3"
-	"github.com/giantswarm/clustertest/v3/pkg/client"
-	"github.com/giantswarm/clustertest/v3/pkg/env"
-	"github.com/giantswarm/clustertest/v3/pkg/logger"
-	"github.com/giantswarm/clustertest/v3/pkg/utils"
-	"github.com/giantswarm/clustertest/v3/pkg/wait"
+	cb "github.com/giantswarm/cluster-standup-teardown/v5/pkg/clusterbuilder"
+	"github.com/giantswarm/cluster-standup-teardown/v5/pkg/standup"
+	"github.com/giantswarm/clustertest/v4"
+	"github.com/giantswarm/clustertest/v4/pkg/client"
+	"github.com/giantswarm/clustertest/v4/pkg/env"
+	"github.com/giantswarm/clustertest/v4/pkg/logger"
+	"github.com/giantswarm/clustertest/v4/pkg/utils"
+	"github.com/giantswarm/clustertest/v4/pkg/wait"
 	corev1 "k8s.io/api/core/v1"
 	apierror "k8s.io/apimachinery/pkg/api/errors"
-	capi "sigs.k8s.io/cluster-api/api/v1beta1"
+	capi "sigs.k8s.io/cluster-api/api/core/v1beta2"
 	cr "sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/giantswarm/cluster-test-suites/v5/internal/state"
@@ -146,13 +146,17 @@ func Setup(isUpgrade bool, clusterBuilder cb.ClusterBuilder, clusterReadyFns ...
 					if err != nil {
 						logger.Log("Failed to get Cluster CR: %v", err)
 					} else {
+						failureReason := ""
+						failureMessage := ""
+						if cl.Status.Deprecated != nil && cl.Status.Deprecated.V1Beta1 != nil {
+							failureReason = string(ptr.Deref(cl.Status.Deprecated.V1Beta1.FailureReason, ""))
+							failureMessage = ptr.Deref(cl.Status.Deprecated.V1Beta1.FailureMessage, "")
+						}
 						logger.Log(
-							"Cluster status: Phase='%s', InfrastructureReady='%t', ControlPlaneReady='%t', FailureReason='%s', FailureMessage='%s'",
+							"Cluster status: Phase='%s', FailureReason='%s', FailureMessage='%s'",
 							cl.Status.Phase,
-							cl.Status.InfrastructureReady,
-							cl.Status.ControlPlaneReady,
-							ptr.Deref(cl.Status.FailureReason, ""),  //nolint:staticcheck // Ignore SA1019 this field is marked as deprecated.
-							ptr.Deref(cl.Status.FailureMessage, ""), //nolint:staticcheck // Ignore SA1019 this field is marked as deprecated.
+							failureReason,
+							failureMessage,
 						)
 
 						for _, condition := range cl.Status.Conditions {