+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Migrate from CAPI v1beta1 to v1beta2 (`sigs.k8s.io/cluster-api` v1.12.3) and bump `clustertest` from v3 to v4.0.0.
+- Bump `cluster-standup-teardown` from v4 to v5.0.0.
+
 ## [5.0.0] - 2026-03-17
 
 ### Changed